requirements.txt
python-vlc==3.0.18121
windows-curses==2.3.1
